update names and add fast paths

+++ python/cudf/cudf/core/column/column.py
@@ -1172,6 +1172,7 @@ class ColumnBase(Column, Serializable, BinaryOperand, Reducible):
         return result
 
     def astype(self, dtype: Dtype, copy: bool = False) -> ColumnBase:
+        # import pdb;pdb.set_trace()
         if len(self) == 0:
             dtype = cudf.dtype(dtype)
             if self.dtype == dtype:
@@ -1251,7 +1252,7 @@ class ColumnBase(Column, Serializable, BinaryOperand, Reducible):
             )
 
         # Categories must be unique and sorted in ascending order.
-        cats = self.unique().sort_values().astype(self.dtype)
+        cats = self.unique().sort_values()  # .astype(self.dtype)
         label_dtype = min_unsigned_type(len(cats))
         labels = self._label_encoding(
             cats=cats, dtype=label_dtype, na_sentinel=cudf.Scalar(1)
